Add accessibility method to `View` and move text edit logic
