+++ crates/vizia_core/src/events/event_handler.rs
@@ -1,4 +1,6 @@
-use crate::prelude::*;
+use accesskit::{NodeBuilder, NodeId};
+
+use crate::{context::AccessContext, prelude::*};
 
 use std::any::Any;
 
@@ -10,6 +12,13 @@ pub trait ViewHandler: Any {
     fn event(&mut self, cx: &mut EventContext, event: &mut Event);
 
     fn draw(&self, cx: &mut DrawContext, canvas: &mut Canvas);
+
+    fn accessibility(
+        &self,
+        cx: &mut AccessContext,
+        node_builder: &mut NodeBuilder,
+        children: &mut Vec<(NodeId, NodeBuilder)>,
+    );
 
     fn as_any_ref(&self) -> &dyn Any;
